feature: Created a prototype for the Loop Fusion pass.
The pass currently consists of empty functions and data structures, which this project will further build upon in the future.
Also added the CMakeLists.txt file for the pass.
The pass is currently called upon with "loopfusion" as there already exists a loop fusion pass in LLVM that is called with "loop-fusion". This will possibly be changed in the future, should the need arise.