New: Added the skeleton of a new function that should compare loop trip counts
Added a bool function IdenticalTripCounts that will, in the future, return true if the trip counts of 2 different loops are the same or false if that is not the case.
The function CanFuseLoops currently returns the result of that function, but will return the conjunction of more checks once they are implemented.