Improve: Implemented basic loop trip count comparison
Implemented the functionallity of the SameTripCounts() function that "should" in theory return the comparison of trip counts.
Unfortunately, the trip counts that are currently being printed out are "COULDNOTCOMPUTE" for all loops, signaling that there is deeper issues that need to be fixed before this function can work.
Most likely, the loops it checks must be simplified somehow, which should be done in the future.