Improve: Added basic loop bound comparison
Implemented a loop bound comparison in the haveSameTripCounts() function that checks if both loop bounds are a constant and if their values are the same. Otherwise, currently it only returns false.
A loop counter check and a check in case the bounds are variables should be implemented in the future.