Improve: Implemented check if both non-constant loop bounds point to the same object
Added an unordered_map and its mapping function to map all variables in the code, so we can now check if two IR variables point to the same thing.
Using the former, implemented comparisons of non-constant loop bounds in the haveSameTripCounts() function.
Now all that is left is to implement a check for the beginning counter values.