Improve: Finalized trip count check to compare loop latch operations and values
Function HaveSameTripCount() now also compares loop latch values and if the latches have the same binary operation.
Just now realized the Trip count, beginning value and latch values could all be checked in separate functions to avoid spaghetti code, and the code should be refactored in the future, if there is time for it.