New: Added basic loop dependency checks
Implemented function areDependant() that checks if two loops depend on one another, in which case they shouldn't be fused.
Currently, the function only checks if there are load or store instructions in the bodies of the loops, which change the same variable.
Also updated the FunctionCandidate class to include a vector of all load/store instruction variables, with it's getter and setter.