Improve: Improved loop dependancy check by separating read and write instructions
Previously, two variables in two read instructions would be deemed dependant, which was wrong.
Now we check for loop body variable if it exists in a read or write instruction. Two writes and a read and a write are incompatible, making two loops dependant. Two read instructions keep them independant.