Improve: Added dependancy checks for loop headers, which previously didn't happen
When checking for loop dependancy, now the increment/iterator is kept as a write variable, while loop bound is kept as a read variable, making some mistakenly fusable loops now not fusable.
Previously, if an iterator was declared before the loops and used in both loops, it would return the loops as independant, which would be true in most cases, but because of the way loop fusion and the LLVM IR are structured, is false.
Also updated documentation for an unfuseable loop example.